Refactor popup dialog to use Hass Dialog Manager.
