+++ custom_components/browser_mod/services.yaml
@@ -197,12 +197,11 @@ popup:
       description: Hide timeout progress bar
       selector:
         boolean:
-    allow_nested_more_info:
-      name: Allow nested more-info dialogs
-      description: Allow nested more-info dialogs to be opened from this popup
-      default: true
+    popup_dialog_tag:
+      name: Popup dialog tag
+      description: Dialog tag for the popup
       selector:
-        boolean:
+        text:
     style:
       name: Styles
       description: CSS code to apply to the popup window
@@ -268,6 +267,17 @@ close_popup:
           multiple: true
           filter:
             domain: "person"
+    all:
+      name: All popups
+      description: Close all popups on the browser
+      default: false
+      selector:
+        boolean:
+    popup_dialog_tag:
+      name: Popup dialog tag
+      description: Dialog tag for the popup to close
+      selector:
+        text:
 
 notification:
   name: notification
