+++ js/plugin/frontend-settings.ts
@@ -346,9 +346,9 @@ export const AutoSettingsMixin = (SuperClass) => {
               const evShowDialog = new CustomEvent("show-dialog", { bubbles: true, composed: true, detail: { browser_mod_continue: true, ...ev.detail } })
               ev.stopPropagation();
               window.browser_mod?.showPopup(
-                'Edit sidebar',
-                'Browser Mod is installed. It is recommend that you use Browser Mod Frontend Settings to manage sidebar settings.',
                 {
+                  title: 'Edit sidebar',
+                  content: 'Browser Mod is installed. It is recommend that you use Browser Mod Frontend Settings to manage sidebar settings.',
                   right_button: "Continue",
                   right_button_action: () => { main.dispatchEvent(evShowDialog) },
                   left_button: "Edit with Browser Mod",
