+++ js/plugin/main.ts
@@ -14,6 +14,7 @@ import "./popups";
 import { PopupMixin } from "./popups";
 import pjson from "../../package.json";
 import "./popup-card";
+import "./popup-dialog"
 import { AutoSettingsMixin } from "./frontend-settings";
 import { BrowserIDMixin } from "./browserID";
 import { VersionMixin } from "./version.js";
