+++ js/plugin/panel.ts
@@ -122,7 +122,7 @@ export const PanelStateMixin = (SuperClass) => {
               ...panelAttributes,
               ...viewAttributes,
               fullUrlPath: fullUrlPath,
-              popupOpen: window.browser_mod?._popupEl?.open || false,
+              popupOpen: window.browser_mod?.popupState
             }
           }
         });
